Reduced duplication between geo and cartesian test
There were 16 tests shared between the geo and cartesian tests.
These are moved into a base class. The reordering of dependencies
required that we move 3 common dateline tests into a utility class.
