+++ x-pack/plugin/spatial/src/internalClusterTest/java/org/elasticsearch/xpack/spatial/search/CartesianShapeQueryTests.java
@@ -13,7 +13,7 @@ import org.elasticsearch.xpack.spatial.LocalStateSpatialPlugin;
 import java.util.Collection;
 import java.util.Collections;
 
-public class CartesianShapeWithDocValuesQueryTests extends CartesianShapeQueryTestCase {
+public class CartesianShapeQueryTests extends CartesianShapeQueryTestCase {
 
     @Override
     protected Collection<Class<? extends Plugin>> getPlugins() {
